feat: add bet amount object, add betting logic in Player
