+++ src/main/java/blackjack/domain/participant/Player.java
@@ -5,6 +5,7 @@ import blackjack.domain.Result;
 public class Player extends Participant {
 
     private final Name name;
+    private BetAmount betAmount;
 
     public Player(final Name name) {
         this.name = name;
@@ -12,6 +13,10 @@ public class Player extends Participant {
 
     public Result judgeByDealerState(final Dealer dealer) {
         return this.state.calculatePlayerResult(dealer.state);
+    }
+
+    public void betting(final int betAmount) {
+        this.betAmount = BetAmount.initBetting(betAmount);
     }
 
     public String getName() {
