+++ src/test/java/blackjack/domain/participant/BetAmountTest.java
@@ -1,6 +1,7 @@
-package blackjack.domain.participant.state;
+package blackjack.domain.participant;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -24,14 +25,19 @@ class BetAmountTest {
     @Test
     @DisplayName("블랙잭에 따른 결과로 곱해진 값을 잘 반환하는지 테스트")
     void testBlackjackBetAmountMultiply() {
-        this.betAmount = this.betAmount.multiply(1.5);
-        assertThat(this.betAmount.getValue()).isEqualTo(1_500);
+        assertThat(this.betAmount.multiply(1.5)).isEqualTo(1_500);
     }
 
     @Test
     @DisplayName("버스트 혹은 패배시 곱해진 값을 잘 반환하는지 테스트")
     void testBustOrLostBetAmountMultiply() {
-        this.betAmount = this.betAmount.multiply(-1);
-        assertThat(this.betAmount.getValue()).isEqualTo(-1_000);
+        assertThat(this.betAmount.multiply(-1)).isEqualTo(-1_000);
+    }
+
+    @Test
+    @DisplayName("초기 베팅 금액이 0보다 작을 경우 예외처리")
+    void testNegativeException() {
+        assertThatThrownBy(() -> new BetAmount(-3_000))
+            .isInstanceOf(IllegalArgumentException.class);
     }
 }
